Share all_initial_forms between commands
