+++ pombola/south_africa/management/commands/south_africa_update_constituency_offices.py
@@ -31,7 +31,10 @@ from pombola.core.models import (OrganisationKind, Organisation, Place, PlaceKin
                          InformationSource)
 
 from mapit.models import Generation, Area, Code
-from ..helpers import fix_municipality_name, LocationNotFound
+from ..helpers import (
+    fix_municipality_name, LocationNotFound,
+    all_initial_forms
+)
 
 organisation_content_type = ContentType.objects.get_for_model(Organisation)
 
@@ -95,38 +98,6 @@ def geocode(address_string, geocode_cache=None):
         lat = float(geometry['location']['lat'])
         return lon, lat, geocode_cache
 
-
-def all_initial_forms(name, squash_initials=False):
-    '''Generate all initialized variants of first names
-
-    >>> for name in all_initial_forms('foo Bar baz quux', squash_initials=True):
-    ...     print name
-    foo Bar baz quux
-    f Bar baz quux
-    fB baz quux
-    fBb quux
-
-    >>> for name in all_initial_forms('foo Bar baz quux'):
-    ...     print name
-    foo Bar baz quux
-    f Bar baz quux
-    f B baz quux
-    f B b quux
-    '''
-    names = name.split(' ')
-    n = len(names)
-    if n == 0:
-        yield name
-    for i in range(0, n):
-        if i == 0:
-            yield ' '.join(names)
-            continue
-        initials = [name[0] for name in names[:i]]
-        if squash_initials:
-            result = [''.join(initials)]
-        else:
-            result = initials
-        yield ' '.join(result + names[i:])
 
 # Build an list of tuples of (mangled_mp_name, person_object) for each
 # member of the National Assembly and delegate of the National Coucil
